# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tool-kit API",
        "description": "Апи регистрации",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8008",
            "description": "Основной API-сервер"
        }
    ],
    "paths": {
...
        "/api/admin/get_all_users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Получение данных всех пользователей",
                "operationId": "6974d95d1b5e6cca1fb2483c60eda7fb",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение данных о пользователе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "schema": "GetAllUsersResponse",
                                            "properties": {
                                                "allUsers": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Takeshi Kitano"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "format": "email",
                                                                "example": "takeshi@mail.com"
                                                            },
                                                            "role_id": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "created_at": {
                                                                "type": "timestamp",
                                                                "example": "2025-02-03T05:34:02.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "timestamp",
                                                                "example": "2025-02-03T05:34:02.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нехватка прав для действий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Не хватает прав для действий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильный токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Неправильный токен"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/update_user/{user_id}": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Обновление данных пользователя",
                "operationId": "c17a14371e8454ab1f07b36d0e08f583",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя (необязательный параметр)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "5"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1..."
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "name",
                                            "email",
                                            "password",
                                            "password_confirmation"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Takeshi Kitano"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "takeshi@mail.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "kiwami"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "kiwami"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Обновление прошло успешно"
                                                },
                                                "newData": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Takeshi Kitano"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "takeshi@mail.com"
                                                        },
                                                        "role_id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "created_at": {
                                                            "type": "timestamp",
                                                            "example": "2025-02-03T05:34:02.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "timestamp",
                                                            "example": "2025-02-03T05:34:02.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "email": [
                                                    "The email field is required."
                                                ],
                                                "password": [
                                                    "The password must be at least 3 characters."
                                                ],
                                                "password_confirmation": [
                                                    "The password field confirmation does not match."
                                                ]
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нехватка прав для действий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Не хватает прав для действий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильный токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Неправильный токен"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/delete_user/{userId}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Удаление пользователя",
                "operationId": "b37189fb0ade4bdb0feae27149afca53",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID клиента для удаления",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Пользователь User Name успешно удален"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Проблема с нахождением пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Пользователь не найден"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нехватка прав для действий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Не хватает прав для действий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильный токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Неправильный токен"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}